test(vetting): empty the last _common.h of values, and enforce that it stays empty
test_2d_remaining_common.h was the last shared header holding reference data: six
tables and nine helpers, read by four other files. All of it now sits with the
assertions that read it, and check_test_names.py rejects a reference table declared
in any _common.h, so the rule from the previous commit is enforced rather than
merely written down.

Two functions had to be split rather than moved, because they spanned oracles:
test_2d_morphology_caliper_regression held two regression keys alongside eighteen
imea ones, and test_2d_morphology_chord_stat_regression held eleven alongside
ALLCHORDS_MIN. Those nineteen imea assertions reached their goldens through
remaining2d_caliper_ref_val(), a lookup spanning an imea table and a regression
table -- so a _regression function was resolving imea-vetted values, and the rows'
own target_test=test_2d_morphology_imea.h was never met. They are now
test_2d_morphology_caliper_stats_imea in that file.

The two imea tables are NOT merged, contrary to the literal review request. They
hold the same features measured on two different fixtures and every one of the 12
shared keys differs by roughly 5x -- STAT_FERET_DIAM_MIN is 4.473 on the shape2d
mask and 21.0 on the a=20/b=10 ellipse -- so merging would collide those keys and
silently drop one fixture's goldens. SPEC 6.3.1's _<subject> qualifier covers this:
morphology_2d_imea_shape2d_ref_vals (which also absorbs the three geodetic keys,
same fixture, no overlap) and morphology_2d_imea_ellipse_ref_vals, both in the imea
file, which is what the request was after.

New test_2d_radial_regression.h: the radial table and test_2d_radial_distribution_
regression lived in test_2d_intensity_histogram_regression.h, so a radial_* table
and a radial-named function sat in a file named for another family. The registry
says family=radial; the three rows now target the new file. This is the one call in
this batch that was flagged for review rather than settled, and it is easy to revert
if the answer comes back the other way.

Line endings: LF is the Nyxus standard, so every file this branch touches is now LF.
That is 22 files and ~7.7k lines, all under tests/ -- nothing in src/. It also
reverses two earlier mistakes on this branch: 6feaf91 gave an include the CRLF "its
file uses", and an earlier pass here normalised two mixed files to their dominant
CRLF. Both were the wrong direction. Content is provably untouched: line counts are
unchanged per file and the whitespace-ignoring diff carries no extra lines.

The other ~205 CRLF files in the repo (89 .cpp, 112 .h, 14 .cu, almost all under
src/) are deliberately left alone -- converting them here would bury the review in a
57k-line diff. They want their own mechanical pass, ideally with a .gitattributes
rule so the standard is enforced rather than remembered.

Registry: the three radial rows point at the new file; CONVEX_HULL_AREA and SOLIDITY
had a stale target_test naming a _regression file for a vetted row, repointed to
test_2d_morphology_skimage.h with no code move; the nineteen imea rows drop
test_2d_morphology_regression.h from current_test now that it no longer asserts them.

gtest 732/732 (one new case), pytest 85 passed / 1 skipped, both checkers clean, and
no reference table remains in any _common.h.
